+++ fetchSwapData.ts
@@ -1,4 +1,4 @@
-import { getPortalfiSwap } from "./protalfi";
+import { getPortalfiSwap } from "./portalfi";
 import { getEnsoSwap } from "./enso";
 import { getBarterSwap } from "./barter";
 import { getApprovalAddressForChain,getMinAmountOut } from "./utils";
